move IIIF manifest endpoint
